move __tests__ dir to src
