+++ packages/react-native-reanimated/src/__tests__/animatedProps.test.tsx
