+++ packages/react-native-reanimated/src/__tests__/bezier.test.ts
@@ -1,4 +1,4 @@
-import { Bezier } from '../src/Bezier';
+import { Bezier } from '../Bezier';
 
 // spell-checker:disable
 /*
