+++ packages/react-native-reanimated/src/__tests__/hooks.useAnimatedStyle.test.tsx
@@ -6,7 +6,7 @@ import Animated, {
   useAnimatedStyle,
   useSharedValue,
   withTiming,
-} from '../src';
+} from '..';
 
 jest.useFakeTimers();
