+++ packages/react-native-reanimated/src/__tests__/normalizeColor.test.tsx
@@ -1,4 +1,4 @@
-import { normalizeColor } from '../src/Colors';
+import { normalizeColor } from '../Colors';
 
 describe('Test `normalizeColor` function', () => {
   describe('Only compliant color are accepted', () => {
