+++ packages/react-native-reanimated/src/__tests__/spring.test.tsx
@@ -1,8 +1,5 @@
-import type { SpringAnimation } from '../src/animation/springUtils';
-import {
-  bisectRoot,
-  scaleZetaToMatchClamps,
-} from '../src/animation/springUtils';
+import type { SpringAnimation } from '../animation/springUtils';
+import { bisectRoot, scaleZetaToMatchClamps } from '../animation/springUtils';
 
 describe('Spring utils', () => {
   test.each([
