+++ packages/react-native-reanimated/src/__tests__/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../../../tsconfig.json"
+}
